make desktop drawer collapse instead of hide
this PR changes the desktop drawer UI to match what is in the mockups
(see the linked issue for details). basically currently, when you're on
desktop the drawer hides and shows itself by sliding off to the left,
this changes things so that instead it's 'hidden' form is like a compact
version of itself.

I also changed it so that the drawer defaults to it's closed state on
desktop, so the drawer doesn't take up quite so much space by default.

closes #1952
pr #1977