add option for custom html on URL widgets
This changes the URL widget so that the user can enter custom html embed
code for embedding content from Twitter. This is because standard embeds
from Twitter with embedly aren't so good for some pages, and some users
would like to be able to use publish.twitter.com to make custom embeds
to include on their channels.

Basically, now there are two fields on the embed widget, a URL field and
a 'custom html' field. The URL field works as before. The 'custom html'
field lets you put in some twitter embed code. There's validation to
make sure that the content is only from twitter and also that the user
only fills out the url field or the custom html field, not both.

closes #1951
pr #2004