add an animation for the drawer expand / contract on desktop
This adds a CSS transition to animate the transition between the
expanded and contracted state for the drawer. Should just make it a
little more visually appealing and less jarring.

I also added spacers and adjusted CSS so that there is no longer a
visual jump up and down when the drawer expands and contracts.

closes #1994
closes #1998
pr #2000