remove ANONYMOUS_ACCESS feature flag
This removes the anonymous access feature flag on both the frontend
and the backend, so that the anonymous access behavior is now the
default behavior. We have had the flag switched on a for a while in
production, so it's time to get rid of it.

As part of removing the flag I deleted a large number of cassettes which were
generated for previously-existing parameterized test cases where the
feature flag was turned off and renamed existing cassettes to ensure
that tests were passing for the right reasons.

closes #1763
pr #2011