add channel nav bar to the post detail page
This adds the channel navbar (the thing that shows 'Home', 'About', etc)
to the post detail page.

It looks like when the element was originally added there was an
inconsistency in the design mockups, some showing it on the post detail
page and some not, so that might be why it wasn't already there.

closes #2005
pr #2007