fix drawer animation
small tweak to the drawer animation. I wanted to have a transition on
the .content div as well, so that when the drawer opened or closed the
change in size in the content div would be animated as well. I goofed
though and forgot that we don't set the width property on .content, but
we change it's size by changing the amount of padding-left. This
corrects the transition property on .content so that the change in width
for the content animates nicely along with the drawer opening and
closing.

pr #2012