fix course search page infinite scroll issue
this fixes an issue with the infinite scroll on the course search page.
basically, the issue was that we had two different types of loading
states, one which was triggered by scrolling down and one triggered by
changing the search text or the selected facets. In the case of the
latter we want to reload the whole result UI, clear the old results, and
start from scratch, but in the former case we want to show the
already-fetched results while the infinite scroll component takes care
of the loading indicator at the bottom. But our existing code only dealt
with the latter case, and didn't distinguish the former adequately
(there is more explanation of this on the issue if this is not clear).

So to fix it I basically just added a value, incremental: boolean, to
the state on the course search page component. When we're rendering our
results we can inspect this value to determine whether the most recent
request that was fired off was an incremental reload (i.e. triggered by
the infinite scroll component) or a full reload (i.e. triggered by a
change in the search text, etc). This lets us determine when to show a
full loading UI and when to let the infinite scroll component do its
thing.

I also removed a flaky test in `CreatePostPage_test` which was, as far
as I could tell, redundant anyhow.

closes #1924
closes #1991
pr #2033