persist desktop drawer open / close value to localStorage
this adds redux-localstorage and configures it to persist the value for
ui.showDrawerDesktop to the localstorage. this will persist the value
that the user sets, and when the app loads next the saved value will be
used to set the initial state for the app. this means that if the user
prefers to have the drawer open on desktop it will be 'sticky' between
full page refreshes.

no ticket
pr #2038