install Formik and use it to implement a separate <CommentForm />
this installs Formik and uses it to create a separate <CommentForm />
component. Before we had this integrated thing for both editing posts
and creating comments, and the code was really kind of a mess. I
basically pulled the comment portion of that out into a separate
component, which allowed me to delete a lot of code. The existing
implementation is still there, but just for post editing. The
Formik-based <CommentForm /> is, I think, a lot simpler and more
straightforward, and doesn't require any redux code to work (nice!).

Going forward, I'll will also use Formik to refactor the existing
<EditPostForm /> and for any new forms we end up adding to the app. I
think also as we go along we should all of our current forms to use
Formik, with the goal of eventually deleting our custom forms-in-redux
code.

pr #2054