upgrade a few dependencies
just the pinned versions in yarn.lock, not the version ranges in
package.json. I did this because there was a warning related to caniuse
and browserslist being out of date that was causing our test suite to
fail. we don't depend on those packages directly, so I just ran a `yarn
upgrade` on the packages which depend on them.

pr #2049