add new banner image to the course pages
this adds the updated course home page banner image, as seen in the
mockups linked from the issue (#1997). I also increased the height of
the banner (only on this page, the component is shared with the channel
page) to match the mockups.

note that there are other tweaks to the course banner I haven't made yet.

part of #1997
pr #2059