restyle the course carousel to match new designs
this updates the styling for the titles of the carousels and updates the
controls for the carousels themselves on the course home page to be
in-line with the new designs. Note that this doesn't include re-doing
the cards in the carousel, I was going to do that in one PR but I'm
splitting that part out to keep the PRs small.

In particular, the buttons for the carousel are updated to match the
design. they should also only appear if you can actually move the
carousel - like if you're on the first page of items there shouldn't be
a button on the left. Likewise, if you're at the end there shouldn't be
a button on the right.

part of #1997
pr #2063