refactor CommentTree to use a separate Comment component
this refactors how we display the comments on the post page (and,
incidentally, on a few other pages).

basically, previously we used this large component called a CommentTree,
which takes a tree of comments returned by the comments API (all
comments for a post) and renders it. This works great on the post page,
but is hacky elsewhere. We ended up re-using that component in other
places to display only a single comment, which was silly.

This component also has some crucial flaws, namely, it isn't
self-sufficient. We want to be able to display a comment in a few
different places (on a post, in a profile's contribution feed, on a
moderation page) and allow the user to interact with it in all these
places in the same way. But because the CommentTree expects all the API
callbacks to be passed in from outside we currently have a super messy
thing where we implement similar functions or use HOCs to get them in
different places (e.g. callbacks for upvoting / downvoting, for editing
and then patching a comment, etc).

Anyhow, I refactored this so there are now two components, Comment and
CommentTree. Comment has all the responsibility for correctly rendering
a comment and managing all the interactions that a user can do with it
(edit, reply, remove, delete if it's their own, upvote, downvote) and
CommentTree is now a much more minimal component that has only the
narrow responsibility of correctly traversing the tree and rendering
either a Comment or a 'click to load more comments' placeholder.

That is the broad strokes. Doing this allows us to easily render a
comment elsewhere without needing much external 'scaffolding' to set up
it (API callbacks and the like) and simplifies the implementation of
things like the channel moderation page.

A note on tests: I deleted a bunch of comment-related tests from the
PostPage tests, but I replicated all of those tests in the new test file
for the Comment component. Likewise, all of the tests in the CommentTree
file that were actually testing how the comment was displayed and so
on I copied over to the Comment test file, and just modified them to
get them passing with the new component. Often the tests did not
change beyond changing a wrapper = render() call to { wrapper } =
await render() because I used the redux-query compatible renderer.

closes #1714
pr #3011