fix styling issues with the drawer
this fixes a few styling issues with the drawer.

the first is an issue that croped up with long channel titles. they
could get weirdly squished onto two lines, breaking the alignment.

the second was that the left edge of the channel titles is supposed to
be aligned with the text in the 'Compose' and 'Home' buttons.

I also fixed a weird behavior with the `transition-delay` property I had
set to make it nicer and more consistent.

pr #2085