add initial implementation of favorites
this allows you to add a course or bootcamp as a favorite, on the course
home page or on the search page. it also shows, via a filled-in or
unfilled star, whether a course/bootcamp is currently favorited or not.

to get this working on the course search page I had to fetch all the
favorited items from the rest API and then merge that response with the
search results. this is a sub-optimal solution. in the future we're
going to refactor to an elasticsearch-only solution for this part of the
UI, but for now this should work alright.

closes #2091
pr #2127