Switch course search to use CourseCard, grid layout
This switches over the course search UI to use a grid (instead of a
vertical list of cards) and to use the new CourseCard component to
render the course results.

I also tweaked the CourseCard slightly so that it now calls the callback
to open the course drawer only when the course cover image or title are
clicked, instead of when the whole card is clicked. I also ported over
the behavior we had previously on the CompactCourseDisplay component,
where clicking on one of the topics on the course would add that topic
to the current selected search facets.

closes #2089
pr #2098