fix bug with the Cell component
this fixes a little issue with the `Cell` component - basically, it uses
a `setState` call and a `resize` event listener to ensure that the
component is responsive. the issue, though, is that currently on master
this event listener is added, and the old one remove, on, uh, every
single render :woman_facepalming:

so basically I fixed it to set the event listener on mount and remove it
on unmount. checkout the docs for the `useEffect` hook for details:
https://reactjs.org/docs/hooks-reference.html#useeffect

basically, the hook looks like this:

```js
useEffect(() => {
  someSideEffect()
}, [ props ])
```

where the second argument (the array) is an array of props which should
re-trigger the effect when they change. so if you pass a `[]` as that
argument that indicates the effect should only run once.

anyhow, that fixes it.

pr #2299