restyle the course-search facets
this restyles the facets to match the newest mockups. in particular,
there is some text styling tweaking, and clicking on the facet titles
now opens and closed the whole facet list.

to do this I also pulled out some unnecessary code, a few things were
being kept in the redux store unnecessarily.

part of #2161
pr #2184