fix an issue with the responsiveness of the search facets
this fixes an issue with the facets on the course search (see the
attached issue) where they were basically not as responsive as they
should be. In particular, if the user clicked a number of facets at the
same time a search request was fired off for each one, which led to some
laggyness in the UI (and probably some unnecessary load on the
elasticsearch). We only care about the last request, so I just debounced
this so we'll wait until the user stops interacting to fire off the
request.

closes #2180
pr #2183