update course home page
this implements some styling tweaks for the course home page. also while
I was fiddling with this component I refactored it to use the
react-redux and redux-query hook APIs, and rewrote the tests to be tests
of the full component.

I also made the width of the course carousel responsive.

closes #2179
pr #2197