update babel-eslint
when I upgraded eslint and some related packages (#2168) there was, at
the time, a bug in the latest version of babel-eslint that broke the
parsing for the no-unused-vars rule. basically, it caused it to have
tons of false positives, marking a lot of stuff that was fine
incorrectly as an unused variable. because of this I had to disable the
rule to get the build to pass.

Anyway, that issue has now been fixed in the latest version of the
package, so I updated to that. this let me re-enable the rule. I also
fixed all of the unused variable errors that we'd accumulated in the
meantime while the rule was turned off, which was a surprisingly large
amount!

pr #2201