add 'grid' style loader to the course search page
this implements a "grid" style loader for the course search page and
also implements a "carousel" style loader for the course home page (to
show while the couse information to populate the carousels is being
fetched).

on the search page we show this loader if the user has the 'grid' view
selected, because the page can reload if the selected facet is changed.

closes #2224
closes #2167