a few styling tweaks for the course search page
this makes a few little tweaks to styling and layout on the course search
page to bring it more in line with the mocks:

- removes the 'course catalog' text from the top banner
- changes the widths of the two columns on the course search page. in
the 'list' view we do 4 columns for the facets and 8 for the results,
and in the 'grid' view we do 3 columns for the facets and 9 for the
results.
- restyles the 'list' view of the card so that the way that the text and
spacing is done is more consistent for learning resources that have a
one or two line title (among other things).

pr #2237