refactor tooltips
This refactors our LoginPopup and SharePopup to be LoginTooltip and
ShareTooltip, and tears out our homemade stuff for showing these popups,
to use a library instead (this one:
https://react-component.github.io/tooltip/)

Basically, now you can just grab the `<LoginTooltip>` and wrap it around
any element you want to prompt to user to log-in when clicked, like this:

```js
<LoginTooltip>
  <div className="authenticated-users-only">
    FUN!
  </div>
</LoginTooltip>
```

and that's it! No messing with the store, no nothing! 🎉 🎊 😊

this also adds a tooltip to the 'favorite' button that we have on the
LearningResourceCard, so that unauthenticated users are prompted to log
in on the course search page (since that page is open to them).

closes #1299
closes #2046
closes #2209