fix the learning resources drawer right-to-left behavior
this fixes up the learning resources drawer. We want to have this
drawer, unlike our normal navigation drawer, come in from the right side
of the screen and live on the right side when it's open. Our current
implementation (i.e. on master before this commit merges) does end up on
the right side, but the animation is all janky and it's just not so
nice.

Anyway, the way to fix this turns out to be just basically to update to
the latest version of the material components package we're using
([rmwc](https://rmwc.io/)), which fixed the issue. So then most of the
work I ended up doing was dealing with that upgrade, mainly because:

- the drawer behavior / classnames / etc is different, so I had to do
  some work to port our current drawer behavior over
- the react-clamp-lines package we were using is, for some strange
  reason, broken in the new drawer, so I replace that by extending our
  existing `TruncatedText` component to support the course drawer use-case
  (showing / hiding the learning resource description text).
- a few miscellaneous other things were necessary because other parts of
  the rmwc package had changed (like the menu, dialog, I think a few other
  random things)

closes #2218
pr #2286