fix UI issue with showing/hiding options on SearchFacet
This fixes a small UI weirdness, which is that currently we show the
show more / show less UI for any search facet where we have more than 5
options. Thus, if there are six options there is a strange case where
clicking "show more" only causes you to see one more option...

Anyhow, this fixes the issue by changing things around so that we now
only show this UI if the list of options is a bit longer, currently 15.
With this change that UI should only be shown for our longer option
lists, namely the subjects, and shouldn't be used for any of the others.

closes #2287
pr #2288