some tweaks to the course search page
this does a series of tweaks to the course search page:

- remove extra dollar sign from course price (only on the card, not in
the drawer)
- values in the card description for platform, topic, etc should be
links to a search for that value (e.g. clicking the 'Math' subject on a
course should take you to a search for the 'Math' subject')
- only the favorite button in the bottom row of the card should have a
'link' cursor
- cards all over the site should have a hover state, where a box-shadow
shows up (see designs)
- course home page search 'magnifying glass' icon should redirect to
search with the text the user has entered (currently this behavior
happens when you hit the enter key but not when you click the little
icon)
- change home page link text from "View All" to "Explore"

this is basically just a few things we wanted to tweak to get the
implementation closer to the designs.

note that when I was working on this I found an issue with how we store
the search state on the course search page - basically, there's a
weirdness with not having a single, consistent source of truth for where
the search options are stored. details here:
https://github.com/mitodl/open-discussions/issues/2333

long story short it just meant that to implement the 'subject' links
described above I had to use normal `<a>` tags instead of using
react-router `<Link>` tags, which means that we have a
somewhat-undesirable full page refresh. I think we can fix this in the
future if we refactor the way we store the search state.

closes #2294
pr #2334