two tweaks to course search facet
this makes basically two small tweaks to the display of the course search
facet, to fix two small issues with its behavior.

the first is that if there are no options for a facet we hide the facet,
so the user doesn't see, for instance, 'Subject Area' with no options to
select.

the second is to fix a small bug where collapsing / expanding the facet
didn't properly hide the 'View Less' / 'View More' text for facets where
that is supposed to show up.

closes #2298
closes #2318
pr #2315