add history to the learning resource drawer
This refactors how we control what is shown in the
LearningResourceDrawer. Previously we basically set an object ID and
object type, and then we'd use those to fetch the right object, and then
we'd show it. This worked for opening and looking at one object, but it
broke down a little when it came to navigating to a new object from the
initial one. The specific problem is that if you click in you can't get
back to the object you we're looking at before without closing the whole
drawer and opening and navigating again.

We need to support this use-case because the user can do this from a
userlist or a learning path.

Anyhow, to fix this I changed it so it works on the basis of a history
object, which we store in the ui reducer. It's just an array, and this
is how it works:

- if the array is empty then nothing is on the history, so don't show
  the drawer
- if there is any item on the history, then show the drawer
- take the last item on the history and use that item ID and object type
  to fetch an object to display in the ExpandedLearningResourceDisplay

so I implemented 3 actions: push, pop, and clear. they do what you'd expect. overall it seems to work pretty well!

closes #2440
pr #2463