/courses/ -> /learn/
this changes all the `/courses` URLs to be `/learn`, so from
`/courses/search` to `/learn/search`, `/learn/lists`, etc.

I also pulled all the relevant routes into a separate new component call
the `LearnRouter`. this means we can also mount components for all those
subpages, so we can render like our dialogs and stuff at that one place
instead of all over the place.

This will hopefully in the future, after refactoring the search state,
let us pull the learn banner out to one place as well.

I also added a redirect from `/courses/` to `/learn`, so a link that is
pointing to the wrong thing will still "work". idk if we'll remove that
in the future, but this will make any existing saved URLs still work.

closes #2360
pr #2471