implement new design for learning resource drawer
This updates the display of learning resources in the drawer to match
our latest designs. I also updated the display of the logo for the LR
platform, and added some assets for the platforms we didn't have a logo
for already.

Basically, this mainly is just restyling things, but there are a few
more major parts:

- I pulled out the styling for the expanded learning resource display
from `course.scss` to a new file, `learning-resource-drawer.scss`, which
has only the styling for the drawer in it.
- I also finally updated a lot of selectors like were of the form
`course-foo` to be just `foo`, since those selectors were originally
written when we were only showing courses and we're now showing
bootcamps and programs and so on.

closes #2466
closes #2253
pr #2486