fix the display of read more / less in the truncated text component
Currently when the `TruncatedText` component is used with the
`showExpansionControls` prop we always show the 'read more' / 'read
less' controls, regardless of whether any content has been hidden by the
`Dotdotdot` component. This changes things so that we only show the
expansion controls in the case where they can, you know, do something.

In order to do this I had to figure out how to tell when the text was
being truncated. Basically, this is being done with a div with
`overflow: hidden` on it, and the `Dotdotdot` component doesn't have a
way to pass in a callback or something to get out whether or not it's
done the truncated.

So what I had to do was get a ref to the div used by `Dotdotdot` and
then look at some of it's properties. Basically if the `scrollHeight`,
which measures the height of the content, is higher than the
`offsetHeight`, which represents the high of the element on screen, then
there is some content which has been hidden. So in this case we show the
expansion controls, and otherwise we can safely hide them. I had to do a
very slight hack to make this work, basically to force our
`TruncatedText` component to always do two render calls so that the
`Dotdotdot` component has completed it's render. But things seem ok!

part of #2466
pr #2484