fix issue with learningResourceSelector function
fixes a small bug with one of our selectors. basically, we sometimes
return a getter wrapped with `_.memoize` to be able to access things in
the store based on prop values. the problem with
`learningResourceSelector` is that we return a memoized function of two
arguments, `objectId` and `objectType`. But `_.memoize` only uses, by
default, the first argument as the cache key for memoization, so if we
did this:

```js
const getter = useSelector(learningResourceSelector)

const obj1 = getter(1, LR_TYPE_VIDEO)
const obj2 = getter(1, LR_TYPE_COURSE)
```

we'd be in a bad situation because `obj1` and `obj2` would be the same
object 😱

this is because when we called `getter` the second time it would only
look at the first argument to figure out whether to return a cached
value or not, and it would return the same object as the first call. To
get around this we can pass another function to generate a custom cache
key, as described here: https://lodash.com/docs/4.17.15#memoize

pr #2489