add ability to share learning resources
this PR sets us up for allowing the user to share learning resources via
a link. There are two components:

- on the learning home page we now look for URL params on startup which,
if present, are used to open the learning resource drawer
- in the learning resource drawer there is now a "share" button, which
opens a little popup (the same one we use for posts and comments) which
has the URL for the course home page with the right URL parameters in
it.

One additional detail is that we hide the "share" button for userlists
or learning paths that are marked "private". In the future we may want
to add a way for users to selectively share their private lists, i.e. by
giving our a "share url" like w/ google docs or by giving permission to
certain users. but for now we don't have backend support for this, so
we're just going to hide the UI element.

closes #2435
pr #2526