refactor course search state to live in the URL
this basically refactors how we store the search state for the course
search page so that the params in the URL are the sole source of truth
for the filters and so on.

right now in master these things are sort of stored in two places - in
the URL bar and in the component state for the CourseSearchPage
component. We have some sort of weird logic for syncing them, and for
getting param values from the url bar on startup.

This change makes it so that a few things (like incremental vs
full-refresh searching) are driven off of values kept in component state
while filters and all that stuff are driven purely off of the URL bar. I
think this makes it simpler overall.

closes #2333
pr #2685