fix podcast card height issues
this just sets a min-height on the div that we use to show offered_by on
the podcast cards to ensure that the podcast card heights are the same,
regardless of whether the offered_by field is filled in or not. without
this cards that have an offered_by are a little shorter than the others.

closes #2883
pr #2920