add FilterableSearchFacet component
this adds a fancier way to deal with a facet that has a lot of options -
basically, we have a text input that filters down the list of options
based on what the user types. so they can type in 'bio' and get things
like 'Biology', 'Biochemistry', 'Biomedical Engineering', etc.

We're using this component now for 'Subject' and 'Offered By'.

I also moved course search-related components into a `components/search`
subdirectory and make a few tweaks to the CourseFilterDrawer and related
components.

closes #2929
closes #2938