fix off-by-one error
this fixes an off-by-one error in how we figure out when to disable the
'Next' button on the pagination controls for lists of learning resources
displayed in the LR drawer.

basically, the button was enabled when it should have been grayed out,
so it was possible to page forward to an empty page.

pr #2995