dependency upgrades, add @reduxjs/toolkit
basically just a) updating webpack and related tooling and b) installing
the reduxjs toolkit and using that to rewrite a few reducers. I upgraded
react and react-dom and related packages that need to be updated at the
same time, but those changes are all non-breaking changes so not much
adjustment was needed.

pr #3008