ensure the AudioPlayer works across the site
this adds the AudioPlayer as a site-wide component on the frontend so
that if you start listening to a podcast on `/podcasts` and then
navigate away the audio player UI will stick around.

I also added an 'x' button in the corner which will dismiss the audio
player by clearing the information from our audio reducer.

closes #3014
pr #3015