refactor ExpandedPostDisplay to be a function-based component
this basically just refactors ExpandedPostDisplay to be a function-based
component instead of a class-based one. This is a step along the road to
a bigger refactor that will remove a lot of excess and redundant code
between ExpandedPostDisplay and CompactPostDisplay and hopefully make
this whole part of the site a lot more straightforward.

But as-is this is basically just move code around.

pr #3039