refactor post voting to be more straightforward
the way we do post voting is sort of a mess right now. We have one
component called a PostUpvoteButton which has responsibility for
rendering the button, but responsibility for declaring the callbacks is
external to it. So anywhere we're display a post we have to pass a
toggleUpvote function in to ExpandedPostDisplay and CompactPostDisplay
so that they can then pass that down to the PostUpvoteButton. Because we
show posts in a variety of places (home page, channel page, post page,
profile contribution feed, search results...) this is wicked messy and
leads to needlessly passing callbacks down the component tree.

So I refactored PostUpvoteButton so that it manages its own callback
without external dependencies. This means that all it needs as a prop is
a Post - simplifying it's API and removing the need to create specific
functions to pass in to it.

this lets us delete a fair bit of boilerplate type code and remove
several redundant tests (well, tests that were testing the same thing
but necessary because we were implementing the same thing in a bunch of
places).

pr #3038