fix audio player drawer padding issue
this basically changes the min-height property set on the .navigation
div in the drawer. we set this in order to force the footer to be down
at the bottom of the page, but the way we were doing it doesn't account
for the height of the audio player. I changed it around so that we're
subtracting the audio player height from that min-height value when
the audio player is open.

closes #3049
pr #3050