fix course search textbox behavior
the aim here is to accomplish for the Open LR search the same thing I
did here: mitodl/hugo-course-publisher#190 for the OCW search -
basically, removing the automatic search behavior and instead requiring
the user to press 'enter' or click the search icon.

I also found a few performance issues on the page and diagnosed those,
adding React.memo to fix.

I also refactored CourseSearchPage from a class-based component to a
functional component. This brings it in line with how the search in OCW
is implemented, which should allow me to pull out the functionality we
want to share between the two as a React hook-based library that we can
maintain separately.

pr #3135