replace node-sass with just sass
This removes the `node-sass` package, which depends on a native (C)
module, and replaces it with `sass`, which is a pure-JS implementation
of Sass (actually compiled from a Dart implementation). This means our
frontend build has one less native code dependency, which is 💯  in my
book! (libsass is also
[deprecated](https://sass-lang.com/blog/libsass-is-deprecated), so
there's that as well).

We had just a few small things that needed to be adjusted in our actual
.scss source files, but it was mostly pretty straightforward to get this
going.

pr #3412