upgrade jsdom
upgrades jsdom so as to pull in the latest version of ws (dependabot
alert here: #3425).

I had to make some modifications to one test file because of a change in
how jsdom handles document.activeElement. Previously, jsdom's
implemention would allow for elements which are not children of
document.body to be focused, which is not standard behavior. In jsdom
v15.2.1 this was changed to be in-line with standard behavior in
browsers. Our test was relying on this nonstandard behavior, so I just
had to make a small adjustment to ensure that in a few tests enzyme's
mount is mounting to a DOM node attached to document.body.

See this issue for details: jsdom/jsdom#2723

pr #3430