Remove custom node_modules resolutions
I think what was happening (see https://github.com/webpack/webpack/issues/6505#issuecomment-366502560)  was that this was forcing all node_module imports to resolve to /node_modules or frontends/discussions/node_modules, including dependencies of dependencies. Whereas the default is hierarchicarchicarchical lookup…

```
So for example:
node_modules (at repo root)
|--babel-runtime
   |--node_modules
      |--core-js			// v2
|--core-js					// v3
|--react-router
	|--node_modules
		|--mini-create-react-context
```

babel-runtime was importing core-js, but the webpack config meant that it was resolving to v3 when it wanted v2. And react-router was looking for min-create-react-context in the top-level node_modules instead of its “personal” node_modules.