move userlist stuff out of fields api
Fields are related to userlists: Field.featured_lists and Field.subfields are userlists (or arrays of userlists). So we had put some userlist type definitions in api/fields.

But userlists are learning resources, and most learning resource stuff is in ol-search-ui. (Should make an ol-learning-resources or ol-course-catalog).

So move userlist stuff to learning-resources / search-ui.